Use new screen template for 404 page
