+++ packages/react-component-playground/src/components/display-screen.less
@@ -21,6 +21,7 @@
     display: flex;
     align-items: center;
     width: 540px;
+    min-height: 132px;
     font-family: @default-font-family;
     font-size: 16px;
     padding: 0 16px;
@@ -71,9 +72,9 @@
     }
 
     .header {
-      margin-top: 24px;
       font-size: 22px;
       font-weight: 300;
+      line-height: 26px;
       opacity: 0.6;
     }
